refactor de envío de reportes para incluir indicadores del nodo central federador
